Make forrest.build.xml able to build projects not in ${basedir}.  IOW, there was an assumption that ${project.home} is ${basedir}, which no longer exists.
This occurred because project.* properties set in forrest.properties are
relative ('src/documentation'), but those set in forrest.build.xml (not defined
in forrest.properties) are made absolute.  When the relative property set from
forrest.properties is later resolved, it is done relative to ${basedir}, not
${project.home}.

The solution implemented here is to have external project.* variables (eg.
project.xdocs-dir) that are relative ('src/documentation/content/xdocs'), and
another set of internal variables (eg. xdocs-dir) that are absolute, resolved
relative to ${project.home}.  The external project.* variables are sourced from
a default-forrest.properties file.


git-svn-id: https://svn.apache.org/repos/asf/xml/forrest/trunk@8902 13f79535-47bb-0310-9956-ffa450edef68