These are the resources used by Chaperon.
This is an exact copy of cocoon-2.1/src/blocks/chaperon/samples.
Please try to keep it in synch.

sitemap.xmap is not used directly but kept here as a reference.

When there will be real blocks, all this will not be needed and
we will simply use the reources in the chaperon block.


git-svn-id: https://svn.apache.org/repos/asf/xml/forrest/trunk@9024 13f79535-47bb-0310-9956-ffa450edef68