Point forrest-css to pelt, as it's more similar.
(will move to it also krysalis-site and make it the default one later today if nobody watching this objects).

Explained better what to do if the skin is no longer present and is aliased. Thanks to Ken Coar for understanding that it was not evident that the site should be totally regenerated and published.

git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@37467 13f79535-47bb-0310-9956-ffa450edef68