Better handling of the two cases of SVG generation and skinconf aggregation. See FOR-229 for background discussion.
- src/core/context/resources/stylesheets/project2text.xsl
Replace to call to document() to get the configuration with a simple
reference to '//skinconf' since the skinconf.xml has been <map:aggregate>'d.

- src/core/context/resources.xmap
Replace the 'generate-resource' <map:resource> with two new <map:resources>:
'generate-aggregate-resource' and 'generate-transformed-resource' to reflect
the difference in using the skinconf.xml in a resource versus transforming
the skinconf.xml into a new resource.
Also replaced the 'pipe-svg2png-resource' <map:resource> with
'pipe-aggregated-svg2png-resource' and 'pipe-transformed-svg2png-resource'
to reflect the difference between the resources that use skinconf.xml and
those that transform it.
Changed the calls to resources that use .svg.xslt to use the
'pipe-transformed-svg2png-resource' and those that use the simple .svg
files to use the 'pipe-aggregated-svg2png-resource'.

- src/core/context/resources/stylesheets/extract-svg.xsl
This is used in the transformation of the aggregated SVG to extract ONLY
the SVG itself. Batik requires that the top element be <svg> and in the
aggregation, we have the <skinconf> and containing element after the
transformation.

Submitted by: Rick Tessner
Issue: FOR-229


git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@43410 13f79535-47bb-0310-9956-ffa450edef68