Applying patch for issue http://issues.cocoondev.org/browse/FOR-377:
Displays link for TXT output.  Modified slightly from the patch attached
to issue FOR-377 (txtpod.png image is used rather than pdfdoc.png).

Also modified the CSS so that the image is displayed with a size of 16x16
size and placed properly as well.

This is only for the tigris skin at the moment.  Other skins will be
updated.  Need to look at the common XSLTs that are included.  I suspect
that the common includes are no longer common. :)



git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@76281 13f79535-47bb-0310-9956-ffa450edef68