Resolves: FOR-403
Copy the {forrest.home}/build/plugins directory to the generated webapp.

This gets plugins working in a deployed webapp.

Not sure this is the best solution.  Suspect this would be better
located somewhere in WEB-INF and perhaps copy the plugins/**/lib/*.jar
to the WEB-INF/lib location as well .... Hmmmm ....



git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@110113 13f79535-47bb-0310-9956-ffa450edef68