Added two new plugins for refactoring views into the core: - structurer - themes
Recent changes to views with using jxtg as core component
made the old view plugins unusable (FOR-675) 
which can not longer stay like this. 

I recommend to rollback:
- org.apache.forrest.plugin.internal.view
- org.apache.forrest.plugin.output.viewHelper.xhtml
to revision -r280939 and then commit them back as views head. 

The changes not related to jx could be readded via patches from svn log.

This will solve FOR-678 for old views plugin and use the new plugins 
to incoperate forrest:views into the core. 

The new plugins will still suffer from FOR-678! 


git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@292072 13f79535-47bb-0310-9956-ffa450edef68