> El lun, 17-10-2005 a las 22:31 +0100, Kevin escribió: > > Thought I'd post my work in progress in trying to do switchable > > alternate css stylesheets. I call it theme switching though not > > sure if that is the correct terminology.
The following branding-theme-switcher.ft is based on Kevin's mail and code. I used his code examples partly (more the idea of it) because it did not 
needed the cookie part. 

FIXME: thorsten -> create a function that reads all possible themes and create dynamic examples. Use the helper-prototype for do it.


git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@329441 13f79535-47bb-0310-9956-ffa450edef68