Adding a StAX implementation of the dispatcher to the whiteboard.
It is very basic and should be seen as proof of concept.

To make the version compile you need to download 
and copy the JSR 173 API from 
http://www.ibiblio.org/maven2/stax/stax-api/1.0/stax-api-1.0.jar 
and copy it to lib/api/. 

After building
./build.sh
try
./dispatch

Right now we using a master structurer and ignore any parameter coming in.
It should be easy to make the structurer configurable like said it is a PoF.

git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@499193 13f79535-47bb-0310-9956-ffa450edef68