The font-family params defined in document-to-fo.xsl were NOT available to this stylesheet, even though it is included after the param definitions.
I have thus added the param definitions here as well, to make the pdf generation in Dispatcher work again, but this is really a cludge that goes against the idea of putting the font-family specifications into params/variables instead of as hard-coded strings.

Anyway, pdf generation in Dispatcher is working again.

git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@673630 13f79535-47bb-0310-9956-ffa450edef68