Passing in font family names as parameters to the stylesheet.
Tested with the standard skins-based seed. The font family can now easily be changed or overrided by a simple user configuration.

TODO: If the user wants a named font family instead of the generic serif, sans-serif and monospace, she still has to change a property within the plugin sitemap, ie change the plugin source. This needs to be changed to using a locationmap lookup instead.

git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@686683 13f79535-47bb-0310-9956-ffa450edef68