Implementation of hook processing.
There is a small downsides in the current implementation: the xpath injection are NOT working well when the injection point is within the main stream of the events. 

It is possible to inject e.g. to /html/head but injecting to /html/body/... will not inject the new events in the principal stream to the location that one w
ould await. The main reason is that the generation of the result document is now as well based on StAX and not any more in DOM.

However to have the same behavior as we have currently in the DispatcherTransformer I do not see any other possibility then using DOM again.

Like with the xml properties I will make the usage configurable since I reckon in 90% of all cases the StAX based calculation is used. For the rest I am stil
l investigating but it seems that DOM is the only possibility because every other possibility that me occur is imitating DOM.



git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@697033 13f79535-47bb-0310-9956-ffa450edef68