Adding an AXIOM implementation of the structurer that supports advanced injection of content.
It actually resolves the corresponding node where one want to inject and not like in the StAX only implementation where we are not able to do so since it 100% based on XMLEvents. 

Needed to extend some AXIOM classes (see WSCOMMONS-389) to create not existing xpath and return a node. In the class AXIOMXPathPatched there is still one fix me to support advanced xpath expression like [@id='example'] which currently is not supported for the generation of new paths.

As soon I fixed this I provide a patch and if this patch makes it into the next release we have to declare the class as deprecated and remove it in the next version of the dispatcher.


git-svn-id: https://svn.apache.org/repos/asf/forrest/trunk@699276 13f79535-47bb-0310-9956-ffa450edef68